Improve render of commit SHAs in submodule diffs
